Delay the area code and zip code loading till just in time
Currently in area gem, the CSV data is loaded when the gem is loaded.
However, because the size of the CSV file is rather large it takes a
while (around 5s in my macbook) to load the data.

This change makes the loading JIT and thus saves a lot of time in the
development environment.

Changes involved:
1. load data during the first method call and use @variable ||=
loading_function to memorize the CSV data.
2. create tests
3. (not directly related to the functionality) ignore the .ruby-version
and ruby-gemset files.